requirements.txt
fastapi
uvicorn
langchain-google-vertexai>=2.0.0
langchain-google-genai
google-cloud-firestore
langchain-core>=1.0.0
google-cloud-logging
mcp[cli]
discord.py
google-cloud-secret-manager
nest_asyncio
pytest
pytest-asyncio
Jinja2
google-cloud-tasks